+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+<a name="1.0.0"></a>
+## [1.0.0](https://github.com/facundofernandez/videojs-social-media/compare/v0.0.2...v0.0.3) (2020-11-02)
+* New configuration structure
+* New customizable options
+* Use of 'feather' icons
+
 <a name="0.0.3"></a>
 ## [0.0.3](https://github.com/facundofernandez/videojs-social-media/compare/v0.0.2...v0.0.3) (2018-07-05)